+++ shc-server/src/main/java/com/shc/shc_server/service/CoordinatorService.java
@@ -1,8 +1,6 @@
 package com.shc.shc_server.service;
 
-import com.shc.shc_server.model.Activity;
 import com.shc.shc_server.model.Coordinator;
-import com.shc.shc_server.repository.ActivityRepository;
 import com.shc.shc_server.repository.CoordinatorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,17 +13,15 @@ public class CoordinatorService {
     @Autowired
     private CoordinatorRepository coordinatorRepository;
 
-    @Autowired
-    private ActivityRepository activityRepository;
-
-    // get all cordinators
+    // get all
     public List<Coordinator> getAllCoordinators() {
         return coordinatorRepository.findAll();
     }
 
-    // get coordinator by id
+    // get by id
     public Coordinator getCoordinatorById(Long id) {
-        return coordinatorRepository.getById(id);
+        return coordinatorRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("not found id: " + id));
     }
 
     // save a new coordinator
@@ -35,21 +31,21 @@ public class CoordinatorService {
 
     // update cordinator already exist
     public Coordinator updateCoordinator(Long id, Coordinator updatedCoordinator) {
-
-        Coordinator existingCoordinator = coordinatorRepository.getById(id);
+        Coordinator existingCoordinator = getCoordinatorById(id);
 
         existingCoordinator.setName(updatedCoordinator.getName());
-        // ...
+        existingCoordinator.setPassword(updatedCoordinator.getPassword());
+        existingCoordinator.setEmail(updatedCoordinator.getEmail());
+        existingCoordinator.setPosition(updatedCoordinator.getPosition());
+
         return coordinatorRepository.save(existingCoordinator);
     }
 
     // delete cordinator
     public void deleteCoordinator(Long id) {
+        if (!coordinatorRepository.existsById(id)) {
+            throw new RuntimeException("not foundid:  " + id);
+        }
         coordinatorRepository.deleteById(id);
-    }
-    
-    // añadir actividad
-    public void addActivity(Activity activity) {
-        activityRepository.save(activity); 
     }
 }
